requirements.txt
Django==5.0.2
git+https://github.com/openai/whisper.git
pyannote.audio==3.1.1
transformers==4.37.2
torch==2.2.0
python-dotenv==1.0.1
djangorestframework==3.14.0
ffmpeg-python==0.2.0 